+++ cgbeacon2/server/blueprints/api_v1/controllers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import logging
+from os.path import exists
 from typing import Union
 
 from cgbeacon2.constants import (
@@ -58,9 +59,15 @@ def validate_add_data(req) -> Union[None, str]:
     if dataset is None:
         return "Invalid request. Please specify a valid dataset ID"
 
-    vcf_samples = get_vcf_samples(req_data.get("vcf_path"))
+    vcf_path = req_data.get("vcf_path")
+
+    # Check that provided VCF file exists
+    if exists(vcf_path) is False:
+        return f"VCF file was not found at the provided path:{vcf_path}"
+
+    vcf_samples = get_vcf_samples(vcf_path)
     if not vcf_samples:
-        return "Error extracting info from VCF file, please check path to VCF"
+        return f"Samples {vcf_samples} were not found in VCF files"
 
     samples = req_data.get("samples", [])
     if overlapping_samples(vcf_samples, samples) is False: